feat(hexagone-web-feature-extractor): switch output from DOCX to Markdown
Replace generate-docx.js with generate-md.js to produce a clean Markdown
file instead of a Word document. Removes the docx npm dependency entirely
— the script now uses only Node.js built-in modules.
